Shove all the RLP code into the source file.
